requirements.txt
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
orjson==3.10.0
pycountry==23.12.11
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sniffio==1.3.1
starlette==0.37.2
starlette-decoRouter==1.0.0
StringGenerator==0.4.4
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
